refactor: mv conv to its own mod
