+++ machine_learning/src/arch/layers/conv/mod.rs
@@ -1,0 +1,7 @@
+mod conv2d;
+mod convolver;
+#[cfg(test)]
+mod test_conv2d;
+
+pub use conv2d::Conv2d;
+pub(super) use convolver::Convolver;
